+++ opentelemetry-user-events-trace/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Add support for RoleName,RoleInstance population from Resource.
+- Add mapping of well known (OTel Semantic Conventions) attributes to PartB.
 
 ## v0.2.0